add rotation to keys, change key-editor
it is possible to use more than one key for same feature (max 3).
The keys are pre-defined in json config either as a string or list (when
more than one key). If saved it is "optimized". Missing keys in own
config will use default now, so no need to adapt config, when we create
more keys.