subdivision crashes solved
The new strategy for keeping materials simply lost textures, so deleteObjects now can be
done without losing the material itself. The subdivided mesh is a
replacement of the original one.

also added a few lines to enable MacOS to run with gl.glPolygonMode, reason for that problem is not yet clear.